nextcloud/requirements.txt
Flask>=2.2.3
flask-cors>=3.0.10
requests>=2.28.2
paho-mqtt>=2.0.0
python-dotenv>=1.0.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.5